Hide the nassnig.org URLs as requested - closes #444
We have been requested to remove the links to each person's
page on http://www.nassnig.org so that personal URLs can
be added instead.  It would be a shame to lose the URLs to
the official page, so instead create a new kind of contact
('Parliament URL' as well as 'Profile URL'), change the
kind of all existing Profile URL contacts to Parliament URL
contacts, and add some code to hide the Parliament URLs.
This will make it simple to unhide the Parliament URLs in
the future, while still allowing Profile URLs to be added
by hand.