Fix the href of links to particular tabs
The href of each link at the top of the tabs is undefined
on person pages, or an unhelpful '/' on place or organisation
pages.  This didn't stop the tabs from working (which was
dealt with by .click) but would break "Copy link location"
from the context menu, and look confusing in the status bar
when hovering over the link.