Add a script for guessing parent areas in Mzalendo
There aren't exact parent-child relationships between
Federal Constituencies, Sentorial Districts and States
in Nigeria.  However, there is code in Mzalendo which
will display parent and child places, such as in the
"Related Places" tab of each Place page.  To add some
somewhat useful data that will be displayed there
before the launch of the site, this script will
find parent-child relationships between two given types
of place, based on whether the area of the intersection
of the child and parent is greater than 98% of the
area of the child.

This is a squash the commits c6d0f31af6 and 9219fa73f,
the latter of which just fixes a typo.