Create a ParliamentarySession model and associated migrations
This is intended so that Place and Position objects can
have a foreign key to the particular parliamentary session
that they're valid for.  It also associates that session
with a particular generation in MapIt.  (I'm not sure
about this particular choice, and there's a comment in
the code to that effect.  Currently it's unused.)

This commit also changes the Place model to add a foreign
key to ParliamentarySession and add the migration.