Remove useless default parameter for update and created dates
The default values in:

  class ModelBase(models.Model):
      created = models.DateTimeField( auto_now_add=True, default=datetime.datetime.now(), )
      updated = models.DateTimeField( auto_now=True,     default=datetime.datetime.now(), )

... are evaluated when the class is defined, which means that the
created and updated fields would be always be inaccurate but for
the auto_now=True and auto_now_add=True options.  Those options
override any default that's specified, so the only effect that
the defaults have is to introduce confusing datetimes into the
models field of South migrations.