Add a method to return ParliamentarySession objects for a PlaceKind
Each Place can have a ParliamentarySession specified for it.
This commit adds methods to PlaceKind that return all
ParliamentarySession objects that are specified for a Place
of that PlaceKind.

There are two versions, one which returns [None] if no
ParliamentarySession objects were found (e.g. as for
places of the "Country" PlaceKind) to make it more easy
to iterate over the sessions in a template.