Add a command to create parliamentary sessions and migrate places
This command is intended to be called from a datamigration to
make sure that ParliamentarySession objects are created for the
known sessions in Kenya and Nigeria, and make sure that the
PlaceKind and Place objects are updated correctly.

Note this uses settings.COUNTRY_APP to distinguish which country
this instance of Mzalendo is for, so it also adds a note to that
effect to general.yml-example.