Create DateFields from datetime.date objects rather than ISO-8601 strs
Thanks to Edmund von der Burg for spotting this problem and
suggesting the solution.  This initially worked fine, but since
410f2f99f86c84c9a472185a611e1ae2d0d8e6c6 the __unicode__ method
of ParliamentarySession uses the year attribute of start_date
and end_date.  Although creating the fields with ISO-8601 date
strs is fine from the database's point of view (it's coerced to
a PostgreSQL date correctly) the attribute of the object seems
to remain a string when its printed later in the script.