Display very small boundary changes differently
Typically, there are a couple of very large overlaps between
a constituency and the new constituencies in the next boundary
changes.  There are also other very small overlaps which are
most likely bad data at the edges of the boundaries.  We don't
wnat to elide these completely, however, since even small
changes might be signifcant if they happen to be a deliberate
attempt at gerrymandering.  So, as a compromise, for any
changes of less than 1% we list them separately in a note below
the more significant overlaps.