Fix the social-and-tools box on the home page
This fixes a problem introduced in 705a5d04e1b7b65ca49ee71a4d7f97,
which changed the layout of the social media box on object pages,
but broke its layout on the front page.

Thanks to Jedidiah Broadbent (@Jedidiah) for spotting that a
"float: right" was missing for .social-and-tools on the front page.