Add a datamigration to set 'place' for Aspirant President positions
This is important so that we can deal with all aspirants connected
with a place in a consistent way - we walk up the hierarchy of
place to the country level, so it's natural to include presidential
aspirants at that level in the same way.  This datamigration only
makes this change if the COUNTRY_APP is set to 'kenya'.