Replace small maps of child places with a textual listing
This commit could have been squashed into the previous one, since
that introduced most of Jed's design.  However, in the future we
may want to revert this and replace the placeholder map images
with real static maps of each child location, so it's useful to
have this as a separate commit.

This commit also bumps the STATIC_GENERATION_NUMBER so that assets
are recached.