Add an admin command to fix party membership position end_dates
This is a fix for issue #550 - some party membership positions
were set to 2012, so they no longer appear as current
party memberships.  Instead, we should set these to 'future'.

We will shortly be importing the new candidate data, in which
lots of these will be ended and new party positions will be
created for a new party.