Add an AlternativePersonName model to replace Person.other_names
The motivation for this change is that keeping alternative
names as a newline separated list in a single column doesn't
fit the relational model well, and makes it awkward to query
on.

This commit also replaces name-related methods with ones that
use alternative_names, and updates the tests that used to
relate to other_names.