Fix the old migration 0038_set_country_for_aspirant_presidents
In South datamigrations you have to refer to models with
orm.ParticularModel instead of just ParticularModel, so that
the frozen version of the model is used instead.  This wasn't
a problem when this migration was run, but may be in the
future if the models that it refers to change.