Fix the old migration 0037_migrate_data_for_parliamentary_sessions
South datamigrations have to use the frozen models via the orm
parameter in the forwards and backwards migrations.  When doing
the migration in an admin command, this wasn't the case - it
was importing whatever the models were in the current models.py.
Instead, remove the admin command and use the code in the
datamigration itself, referring to the models via the orm
object.