Add a script to remove some duplicate aspirants
Because some name matches that were missed, we've ended up with
some duplicate aspirants.  In a previous commit I introduced a
script that adds some alternative names for the correct aspirants;
this script checks within each race for any candidate's name that
matches an alternative name of another - in those cases the
other aspirant is removed.