Make Person and Position's __unicode__ methods consistent
When tracking down problems with Person and Position objects
it can be confusing that they use different representations
of the person's name.  This commit switches Person to also
use legal_name.  This has the advantage that you should never
cause another query by calling Person's __unicode__ but the
disadvantage that the admin interface will show the legal_name
in various places instead of the (probably better-known)
alternative name that's marked as the one to use.