Replace PIL with Pillow
Pillow is a more actively maintained fork of PIL, and PIL causes
problems on some systems that break sorl-thumbnailer in
difficult-to-figure-out ways.  (This is the "IOError: decoder zip
not available" error.)

Pillow is a drop-in replacement for PIL, except for requiring
a different import statement.  All the uses of PIL in
Mzalendo do fallback to the Pillow import if the PIL version
doesn't work, so this should be fine.

Update: we can't actually use this on mySociety's servers,
so this branch is just left for future reference.  The problem
is that the version of virtualenv on our servers is old enough
that the default is to use system packages; we can't change
this because we rely on some system packages on the server:
installation of GDAL needs to be fixed up manually if you're
installing it in the virtualenv, whereas the Debian
python-gdal package is fine.  So, this commit doesn't work
on our servers because the uninstall would try to remove the
system PIL and refuses to do so.