Add CSV files with election results from The Kenya Gazette
These are the official results of the election from the
Kenya Gazette, provided to me by Jessica Musila.

This is The Kenya Gazette Vol. CXV No. 45, Gazette Notice
No. 3155 - although it's down at the moment, I think this
must be from:

  http://www.parliament.go.ke/plone/senate/business/papers_laid/paper-laid-13-march-2013-vol.-cxv-no.45/at_multi_download/item_files?name=Gazette%20Vol.%2045%20Special.pdf

Thanks to Jedidiah Broadbent for saving the PDF as a
spreadsheet; I've extracted CSV files just for each
table and tidied up the data somewhat.