Add a boundary import script for South Africa
This is designed for data from http://www.demarcation.org.za/

This should probably be rewritten, as described in the
comment at the top of the file, to use a different set of
shapefiles and just load them using mapit_import - this
would cut down the code a lot.  However, I only realized
the shapefiles at each level were available when this
script was virtually finished, so it doesn't seem worth
rewriting it now.