In migrations, don't assume any (Place|Organisation)(Kind)? exists
Migration 0037 makes the assumption (if you have the 'nigeria' or
'kenya' COUNTRY_APP) that certain Organisation, OranisationKind
and PlaceKind objects exist, which it shouldn't do - this means that
if you have COUNTRY_APP set to one of those and migrate from the
start with an empty database, this migration would fail.

This commit changes each 'get' of one of those objects to a
'get_or_create', fixing #644.