Improve the titles of pages for viewing positions
Now the position views can selected based on OrgansiationKind
and Organisation, the title should have more than just the
Position Title.  With this commit, for example, a page showing
all those with PositionTitle "Member" and Organisation "National
Assembly" will be:

   Member of National Assembly

... instead of just:

   Member

Similarly, if you specify an OrganisationKind (e.g. "House") but
not an Organisation, the title will be:

   Member of any House

... instead of just:

   Member