switch-country.py: work around media_root SuspiciousOperation errors
Having media_root as a symlink that can point to one of a number
of different countries' sylinks doesn't work out well, since
on generating thumbnails, the path to write to is resolved completely
and a SuspiciousOperation exception is raised if that path does
not begin with MEDIA_ROOT.

So, instead, just use a sylink media_link.COUNTRY -> media_root
to indicate which country's media_root is currently used.  This
complicates the script, so it's less satisfactory in that respect,
but does allow the testing of uploading files and generating
thumbnails.