Add an option to order positions by the person's legal_name
This is part of the fix for issue #422.

The positions.distinct('place') in finding child places seemed
to be forcing an ORDER BY on the place name, although I would
have thought that calling the distinct method just returned a
new QuerySet rather than affecting positions as a side-effect.
In any case, that call is easy to remove, since we were using
set on the results anyway.