Display postal address for each constituency office.
There's nothing in the data structure to prevent more than one postal
address for an Organisation, so I've set this up to list them all,
with luck there should only every be one per office.

(formatting of the HTML should be tidied up later when we're sure this is what we want).