shinyoureye: add a background image requested by EiE to every page
This is as a result of some email correspondence with EiE who
have requested a background image on every page.  This commit
introduces the original image they requested
(mz-nigeria-background.jpg) and the "watermark" version,
generated by Jedidiah Broadbent, which should be less distracting
when trying to read text overlayed on it
(mz-nigeria-background-watermark.jpg).