Add an Django admin command to merge two people
This replaces a series of manual manual steps which you used to
have to perform if you ended up with duplicate people in the
database, for example after a big data import where names are
ambiguous.

Thanks to Edmund von der Burg for reviewing this code,
providing a fix and suggestions for improvement.