Show local MPs for constituency offices
Currently I'm not sure if this is the best approach to getting the MP
for each office. For starters it causes n+1 queries for the constituency
offices while we get the MP, and I'm also not convinced this is the best
way to get MPs. But for the time being it does what it needs to and is
sufficient to allow styling of this page.