Change haystack INDEX_NAME when running tests
This uses the [same technique sayit uses][1] for detecting when
we're running tests, however instead of switching to the SimpleEngine,
we're overriding the INDEX_NAME setting by appending '_test' to it.

Originally I did try using the SimpleEngine, but it [caused errors][2]
when running the tests. Doing it this way also means our tests run
in an environment that's closer to production.

[1]: https://github.com/mysociety/sayit/blob/master/spoke/settings/base.py#L231
[2]: https://gist.github.com/hecticjeff/11ed0b34b101f87cee3d