Only generate breadcrumb links for actual urls
This adds some extra checks to see if the breadcrumb url can be
resolved, if it can, then we generate the breadcrumbs as normal, if we
can't resolve it then we still display the breadcrumb, but we don't make
it a link.

This also solves the problem of some of the breadcrumbs being relative
rather than absolute links, which cause 404 errors.