Abortive attempt at prefetch_related
In attempt to make some headway on #810, looking at URLs such
as /organisation/anc/

Have identified potential SQL queries with Django Debug Toolbar,
and attempted to add prefetches as per this commit.

As the ANC data is so huge, it makes the debug output slow and
unwieldy, so I have been testing against COPE and smaller parties.
The additional fields are, for example, positions, places, images, etc.

Adding prefetch_related to both ZA and core views does not, on refresh
of page, appear to reduce number or time-taken of queries.