Help latest pip and old setuptools to work together.
The latest version of pip doesn't play nicely with the ancient
version of setuptools which comes with Debian - it requires
setuptools 0.8, and we only have 0.6.

http://stackoverflow.com/questions/20905350/latest-pip-fails-with-requires-setuptools-0-8-for-dist-info

In order to fix this, we use a script get_pip.bash from commonlib
to securely install pip and then update to a nice new setuptools.

It looks like this is going to be fixed in version 1.5.1 of pip:
https://github.com/pypa/pip/issues/1422