Fix autocompletion on the admin's hansard alias page
In the Django admin, jQuery is loaded by default but renamed to
django.jQuery, a name that isn't expected by the autocomplete plugin.
jQuery (as its default names) is also loaded on any page that contains
a MarkItUpWidget, but there's no such widget on the page for editing
hansard alias, so autocompletion fails due to the absence of jQuery.

This commit changes the admin pages to include a working version of
jQuery if 'markitup' isn't contained in the string representation of
the media object in the template context.

An alternative fix would be to switch to using a Javascript library
such as RequireJS to load jQuery on demand if it wasn't already
loaded; it would be worth considering using such a system in the future
to make it easier to manage these complex dependencies and version
incompatabilities.

Fixes #1134.