Add django-debug-toolbar to help with local development
This commit adds the very helpful django-debug-toolbar plugin, which
is enabled by default when DEBUG is True and you're connecting to
the server at localhost.  Matthew Somerville pointed out that the
toolbar can cause page loads to be very slow when you have many
templates used on a single page, so (similarly to SayIt) I've added
a DEBUG_TOOLBAR option in general.yml-example, defaulting to True
which you can change to False to disable it.

This commit doesn't set INTERNAL_IPS, leaving it at the default
value of (), so the toolbar uses '127.0.0.1' and '::1'.  If someone
in the future is keen to use this on their development host, for
example, they can add a config option to set INTERNAL_IPS.  However,
so much information is revealed by the toolbar that this shouldn't
be done without careful consideration.