Sort positions by last name on SA organisation detail.
This sorts the existing list of positions by the last element of the
name after splitting by whitespace. It's not exact, but we don't have
definite name parts to use. The organisation_detail.html template
wasn't using the list of positions from the context before in any case,
so it wasn't sorted at all.