Stop each call to Person.name from generating an extra query
We're frequently presenting lists of people in Pombola, where
we would want to use prefetch_related('alternative_names') to
fetch all alternative names for the people in a single query.
However, using filter in the name method causes a new queryset
to be generated and evaluated for each call to name regardless
of whether prefetch_related was used, which can vastly inflate
the number of queries needed to display lists of people.