Pin mysociety dependencies to specific commits
The changes here pin the mySociety dependencies we get from git repos to specific commits in those repos, to avoid breaking changes in the future biting us on deploys. Because we previously used the -e switch to install them as source packages, it's necessary to remove the folders pip put into our virtualenv first, otherwise it'll ignore our new requirements. For #701