Improve speech search result template for south africa
I've made use of the ancestors of the speech section to show exactly
where the speech being return is (eg: Hansard, etc) and included the
speaker name. I've also made it a bigger and more obvious link by
wrapping the whole matched excerpt in an anchor, rather than just a
little bit.

Some styling to make this look a bit better is still needed, but
hopefully this goes some way towards #1163