Make sure Javascript expansion of Hansard section expands the right one
It seems that t.grouper isn't neccessarily unique with a page, so we
need to add something to the element ID to disambiguate it.  This commit
appends a dash and then the ID of the first contained section; if there
are no contained sections then it will just append a dash.