Fix the calculation of overlap between polling unit and larger areas
Previously this was taking the smallest polygon out of any of those
that make up either area and using that as the denominator of the
calculation.  This doesn't seem right; what we want is to see how
much of the LGA area that contains the polling unit intersects with
the larger area.