An old migration means we still need django_sluggable.
At some point in the past za-hansard had a django-sluggable field
on one of its models and this made it into a migration. Even though
the field in question has now gone, in order to run all the migrations
you still need to have django_sluggable installed.

You'd hope that pip would pull this in from the requirements.txt of sayit,
but sadly it doesn't delve into the requirements.txt files of requirements.