People related to a place must be currently related, not historically.
Change related_people method on a Place so that it only returns current
people. At some point in the future we may find ourselves wanting to
provide historic people as an option to this.

While I'm here, move the distinctness to the position subquery, allowing
ordering by '-position__sorting_end_date_high', which was impossible before.