Make Place.related_people include positions, and take a position filter
The people who are returned by Place.related_people were determined by
their positions, but those positions weren't easily accessible in the
template that displays them; this was confusing because you wouldn't
see *why* those people were being displayed.

In addition, each country has different criteria for considering
positions important (e.g. in some countries it might only be MPs,
in some it might be senators, and anyone who's a member of a particular
house).  To deal with this, related_people now takes a position filter
which defaults to a country-specific function.