Make the Place overview page only show relevant representatives
There was a strange inconsistency where the People sub-page for
a Place showed:

  - people who had a current significant position associated with
    the place

  - and people who had a current significant position associated
    with any child place of the place

... whereas the overview page would show people who had any
political position associated with the place (i.e. it would
include extra people who had positions in a particular council
for a place, or local party heads for the place, etc.)  As well
as the inconsistency, it wouldn't be immediately clear why those
people were appearing on the overview page, and they weren't
really "representing" the page as the section heading suggested.

This commit changes the overview page to also use the
related_people method to find those to show on the front page.