Add settings modules for countries and country-specific tests
With these changes, you can now run the tests for a specific
country with, e.g.:

   ./manage.py test --settings=pombola.settings.tests_south_africa
   ./manage.py test --settings=pombola.settings.tests_kenya

... regardless of what COUNTRY_APP is set to in your general.yml

(The non-country-specific tests can still be run with:

   ./manage.py test --settings=pombola.settings.tests )

It also adds settings modules specific to particular countries,
and manage.py has been changed so that when you run, e.g.:

   ./manage.py runserver

... and COUNTRY_APP in general.yml is set to 'nigeria', the
settings module used would be pombola.settings.nigeria.
wsgi.py has similarly been changed to use the settings module
specified by COUNTRY_APP.

This commit incidentally also makes it possible to run pombola
with COUNTRY_APP set to None, a benefit of this being that it's
a useful step to having a generic theme, as with Alaveteli, that
can be overridden and elborated by other implementers.

A further advantage of this change, beyond being able to run
country-specific tests regardless of your COUNTRY_APP settings
in general.yml, is that various country-specific setttings no
longer have to be specified in general.yml, such as:

  OPTIONAL_APPS
  BLOG_RSS_FEED
  BREADCRUMB_URL_NAME_MAPPINGS
  TWITTER_USERNAME
  TWITTER_WIDGET_ID
  MAP_BOUNDING_BOX_NORTH = -22.06
  MAP_BOUNDING_BOX_EAST = 32.95
  MAP_BOUNDING_BOX_SOUTH = -35.00
  MAP_BOUNDING_BOX_WEST = 16.30

... since they're determined by the country, and now kept in
pombola.settings.[COUNTRY_APP]_base.