Make "./manage.py test" work out-of-the-box
We recommend now that people use the ./run-tests script to run
tests, since that will run the non-country-specific tests as
well as those specific to each country.  However, Duncan Parkes
pointed out that lots of people will expect, when new to a
project, for "./manage.py test" to run properly.  (At the moment
it won't because (a) it will try to run country-specific tests
and (b) it will try to run tests that depend on particular Django
applications even though they may not be in INSTALLED_APPS for
the current country. (a) could be worked around pretty easily
with nose attributes (-a !country in NOSE_ARGS in base.py) but
(b) is more difficult.

This commit changes the manage.py script to spot when someone
is running ./manage.py test without specifying a settings module,
and in that case will make the default settings module
pombola.settings.tests, which runs all the non-country-specific
tests with an appropriate INSTALLED_APPS.