Stop using deprecated django.utils.simplejson.
Replace import of simplejson from django.utils with a straight
import of simplejson.

Since we already had simplejson installed from requirements.txt,
this is the version which would have been used anyway by the import
from django.utils. See

https://docs.djangoproject.com/en/1.5/releases/1.5/#system-version-of-simplejson-no-longer-used

for details.

Just to note, we currently have a mix of using simplejson 3.3.2 from PyPI
and json from the standard library, which is based on simplejson 2.0.9
according to the Django 1.5 release notes. It would be nice to note where
we really need simplejson rather than just json.