Import slugify from django.utils.text rather than django.templates.
Hopefully this will have no effect at all, but it's slightly neater.
The slugify function is now available from django.utils.text - this
gives you just the function rather than one decorated as a template
filter, and is now the recommended place to get it.

I've also reordered a few other imports to make them more consistent.

A couple of tests needed updating as the argument to slugify should always
be in unicode.