Redirect to source URL for committee minutes with public set to False
The default get_queryset for SpeechView will only return Speech
objects where public is True.  This creates problems for South
Africa, since we want to be able to follow results from a private
committee minutes through to the original URL.  So, this commit
overrides get_queryset, adds the redirection behaviour and tests
for that.

Fixes #1302.