Allow django-ajax-selects to ensure its JS dependencies are present
Autocompletion of people in the admin interface, e.g. at:

   /admin/hansard/alias/add/

... was broken.  There were two places where ajax_select.js was being
loaded, but our manually added inclusion was 404 anyway, so I've
removed that.  The automatically added inclusion on the other hand, was
failing because window.jQuery was undefined, since the while the admin
has included jQuery, it's been renamed to django.jQuery.  The simplest
solution to this is to remove the setting AJAX_SELECT_BOOTSTRAP = False
(the default being True).  This means that django-ajax-selects will
load jQuery and jQuery UI at the versions it requires if they aren't
already loaded.

This seems to fix #1305 without breaking django-markitup, which was
the concern with the previous related change 3f65d0fe8e1e8.