Restyle the SayIt-based presentation of speeches in Pombola ZA
After switching to SayIt master, the presentation of the speeches
needed some work.  (In some ways it was accidentally improved, e.g.
in that the aspect ratio of thumbnail images of politicians was no
longer broken, but there was still plenty of room for improvement.

This adds some of Martin Wright's CSS work in the main SayIt
respository for style the section pages, and adds a simplified
version of the individual speech pages.

Overall, I think this is an improvement, but there's still plenty
of scope for improvement.