Optionally let the ZA _breadcrumbs partial include the section itself
When used from a section view, it doesn't make sense to include the
section itself in the breadcrumbs.  However, from a speech view, we
might want to use this partial passing the section the speech is in,
in which case you would want that section to be rendered.  This adds
an includethis option so that you can get this behaviour.