Shift sub_page to view initialisation arguments.
We were always setting 'sub_page' in the extra arguments passed
as if they had been found in the URL. Instead of setting it here
every time a url is accessed, let's set it at setting up urls
time in the kwargs of as_view. This means that we can then
access all of those things directly as attributes of the view
rather than as values in self.kwargs, and it also means that
we can set class a class attribute of sub_page in subclasses
rather than doing it in the urls.py file, allowing the splitting
of, for instance SAPersonDetailSub into subclasses.