ZA: in the Rep Locator, redirect straight to a unique place result
If there's only a single result for a place search in the rep locator
it doesn't make sense to present that as a choice - we might
as well redirect straight to the result.

This is South Africa specific because the 'latlon' view is only
available with COUNTRY_APP set to 'south_africa'.

Fixes #1314