Fix one of the duplicate entries for MB Sesele in the Popolo JSON
There were two people in south-africa-popolo.json with the name "MB
Sesele"; however, one of them has an email address tsopoo@fsl.gov.za
which suggests that the name for this entry should be M.A Tsopo (who
otherwise has no entry in the file). In addition, that entry has
the role "Speaker for Free State", which seems to correspond with her
role as described in recent news reports.

So, this commit changes the name of that entry to M.A. Tsopo and adds
her alternative name Ouma Tsopo.