Factor out code to wrap test snippets in <li>s with internal separators
This means that the main loop in the breadcrumbs filter only has to
worry about producing a list of the HTML for each breadcrumb element,
and not inserting separators.

Note that this also changes one of the tests: one of the tests had
inconsistent whitespace (two spaces instead of one) so this fixes that
too.