Fix a potential (but probably difficult to do) XSS attack
Text in the URL was being emitted back into the page (after being
somewhat transformed) by the breadcrumb generation code.  For example,
if you included '<b>' in the URL it would make some part of the text
of the returned page in bold text.  This isn't obviously exploitable,
since the transformations (splitting on slashes, title-casing, etc.)
make it difficult either to include a script from a URL on another
site or include some malicious Javascript directly in a <script> tag.

However, this is by no means assured, and instead we should correctly
escape the text used in breadcrumbs before including them in the
returned page - this commit introduces that change.