Fix breadcrumbs when there are non-ASCII characters in the path
The URL that is passed to the breadcrumb template tag is typically
a unicode, and may contain non-ASCII characters.  (It's either passed
a literal ASCII string, or a unicode returned from
request.get_full_path, which will have decoded any percent-encoded
UTF-8 sequences in the URL and returned the path as a unicode.)

Such paths would cause an error email - in particular we were getting
some of those because the 404 page for
/person/joshua-odongo-onono/%E2%80%8E (with a stray U+200E LEFT-TO-RIGHT
MARK at the end) includes the breadcrumb generated from that path.

To fix that, this commit makes us handle the path and path elements
as Unicode until the point where we generate HTML.  When they are
used to generate the breadcrumb HTML, any unicode characters are
replaced by Unicode character entities or percent-encoded UTF-8
sequenced, depending on the context.

Fixes #1333