Fix traceback in django admin when saving a person.
There's a backwards incompatible change in Django 1.6 not noted in the
release notes. The decision as to whether or not to render a primary key
as a hidden field in inline admin is now made by a new function
needs_explicit_pk_field rather than has_autofield, so we need to
change our template to use that.