Switch from using simplejson (deprecated) to json
Using simplejson (the version from Django) gives a deprecation
warning in Django 1.6 and the warning suggests using the core json
module instead.  There seems to be no reason not to be using
the core json module instead of either (a) the Django version or
(b) the PyPi version referred to in requirements.txt.

(If there is some reason to be using simplejson instead, it would
be worth adding a test that demonstrates why.)