Make core_render_template pick an allowed hostname for its fake request
The core_render_template management command is used during deployment
of our live sites to generate a 'down.html' that can be served
statically. This uses a fake request object, but the name of the server
used in that request is now checked strictly against ALLOWED_HOSTS in
Django 1.6, so instead of using the hostname 'fake' use the first entry
from ALLOWED_HOSTS (if it's non-empty).