Update django admin template tabular.html to match current Django.
This patch was arrived at by:

1) Looking at where we started overriding the tabular.html django admin
template in 7c5f52efbc12aaf6adb36020c0197c7370a94cb1 and finding the django
version from requirements.txt (this was 1.3.5).

2) Checking our changes since we created this file, which are actually all
just things that have changed in the Django version too with different
whitespace.

3) Taking a copies of tabular.html from both Django 1.3.x and
our version in 7c5f52 and producing a patch.

4) Taking a copy of tabular.html from Django 1.6.2 (the version we're using
at the moment).

5) Manually applying the patch to that file - there were local changes
around it from Django, but it was obvious what needed changing.

This should get us all the changes that Django have made to that file since
we started overriding it, and the change that caused us to override in the first
place.

This should fix #1378 (I hope).