Allow non-inline editing of AlternativePersonName in the admin
We were getting a NoReverseMatch error in the admin when viewing the
admin page for a person who has at least one alternative name when
TEMPLATE_DEBUG = True, since it was trying to get the admin URL for
editing that AlternativePersonName on its own page, for the "Edit in
admin" link.  To workaround this, make get_admin_url return False
for AlternativePersonName.  (Thanks to Matthew Somerville for
suggesting this simpler solution.)

This only appeared when TEMPLATE_DEBUG is True, because the tabular
inline template admin/edit_inline/tabular.html was being rendered
from this code in django/template/loader_tags.py:

    try:
        template_name = self.template_name.resolve(context)
        template = get_template(template_name)
        return self.render_template(template, context)
    except:
        if settings.TEMPLATE_DEBUG:
            raise
        return ''