Deal with submissions to either form on the county performance page
Paul has said that we only need to create entries using the existing
feedback mechanism; so this just adds submissions for petition
signatures or comments for the senate to the feedback for these
pages respectively:

  /county-performance/petition
  /county-performance/senate

... marked as non-actionable.

There is some juggling of class-based views required to work with
two forms on a single page, and so that if the form don't pass
validation, the user is redirected to a reduced version of the
page with just the form they were trying to fill.

Note that this doesn't confirm that the supplied email address
actually works (as with the existing feedback mechanism) - for a
serious petitions system one would want to make sure the email
address entered is usable.