Fix malformed file slugs
This is a possible fix for #1304 - there's a migration to
regenerate slugs, with redirection from the old slug.  It removes
the option to choose your own slug in the admin when uploading a
file and just uses slugify on the basename of the uploaded file.

Whether this is a good solution or not depends rather on:

 (a) How much we care about having capital letters in the URLs for files
 (b) Whether people value being able to choose their own slug for a file

I'm inclined to think that this isn't worth it, and we should just not
check the file_archive_file table's slug in the cron job that checks
for malformed slugs.