Don't check slugs in the File model of pombola.file_archive
The slugs for uploaded files can be chosen by the person who uploads it,
and doesn't seem to be important that these filename don't contain
capital letters, for example.  We're getting a lot of emails complaining
about these slugs, so just remove this model from the list that are checked
in core_list_malformed_slugs.

This seems like a better fix for the actual problem described in #1304
(irritating cron emails) than the alternative approach in the
issues/1304-possible-fix branch, which tries to fix them and prevent
people from creating such slugs in the future.